chore(ui): refactor form section
